requirements.txt
numpy
Pillow
smartcrop
